Move 2 file s from /09 to /08
