+++ src/Homework08/Person.java
@@ -1,4 +1,4 @@
-package Homework09;
+package Homework08;
 
 /**
  * класс содержащий в себе информацию о полях и методах.
